[WIP] Attempt №1 at fixing buggy Wolf lunge attack
I've made it so the wolf's collider turns itself into a trigger during
the wolf's lunge, so that it can't collide with the player and stop in
one place.

This produces a new problem, however - when the wolf jumps near a wall,
it will just fall through the world. So, things are still WIP.